requirements.txt
asgiref==3.3.4
Django==3.2.4
dnspython==2.1.0
filelock==3.0.12
idna==3.2
numpy==1.20.3
pandas==1.2.4
py3-validate-email==1.0.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
